requirements.txt
anthropic>=0.42.0
flask>=3.1.0
python-dotenv>=1.0.1
python-docx>=1.1.2
PyPDF2>=3.0.1
markdown>=3.7
gunicorn>=23.0.0
werkzeug>=3.1.0
